chore: move signin and signup to auth domain
