+++ domains/Auth/SignUp/index.js
@@ -11,14 +11,14 @@ import {
 import { ScrollView } from 'react-native-gesture-handler';
 import { Formik } from 'formik';
 import * as yup from 'yup';
-import { retrieveSignUpFunction, retrieveSignInFunction } from '../../services/parse/auth';
+import { retrieveSignUpFunction, retrieveSignInFunction } from '../../../services/parse/auth';
 
-import FormInput from '../../components/FormikFields/FormInput';
-import TermsModal from '../../components/TermsModal';
+import FormInput from '../../../components/FormikFields/FormInput';
+import TermsModal from '../../../components/TermsModal';
 // STYLING
-import theme from '../../modules/theme';
+import theme from '../../../modules/theme';
 
-import I18n from '../../modules/i18n';
+import I18n from '../../../modules/i18n';
 
 const validationSchema = yup.object().shape({
   firstname: yup
@@ -163,8 +163,8 @@ export default function SignUp({ navigation }) {
               {formikProps.isSubmitting ? (
                 <ActivityIndicator />
               ) : (
-                <Button mode="contained" theme={theme} style={styles.submitButton} onPress={formikProps.handleSubmit}>{I18n.t('signUp.submit')}</Button>
-              )}
+                  <Button mode="contained" theme={theme} style={styles.submitButton} onPress={formikProps.handleSubmit}>{I18n.t('signUp.submit')}</Button>
+                )}
 
               <TermsModal visible={visible} hideModal={hideModal} />
             </>
